+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v2.0.19.7
 - Add compatible meta tag for ie browsers
+- Reduce size of carousel videos
 
 ## v2.0.19.6
 - Check if localstorage usage permission is enabled