+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add compatible meta tag for ie browsers
 - Reduce size of carousel videos
 - Fix back button cannot load default video issue
+- Unmute video when user clicks on it
 
 ## v2.0.19.6
 - Check if localstorage usage permission is enabled